# server/static/openapi/merged.json
{
    "servers":  [
                    {
                        "url":  "/"
                    }
                ],
    "paths":  {
...
                  "/round/latest/signed":  {
                                               "get":  {
                                                           "summary":  "Get the latest round with signature",
                                                           "tags":  "rounds",
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/SignedResponse"
                                                                                                                                             },
                                                                                                                                  "examples":  {
                                                                                                                                                   "ed25519":  {
                                                                                                                                                                   "value":  {
                                                                                                                                                                                 "algo":  "ed25519",
                                                                                                                                                                                 "signature_hex":  "abcd...",
                                                                                                                                                                                 "header":  {
                                                                                                                                                                                                "to_height":  18193,
                                                                                                                                                                                                "round":  5,
                                                                                                                                                                                                "root":  "2f1f018e...4bb9",
                                                                                                                                                                                                "tx_count":  16837,
                                                                                                                                                                                                "from_height":  18190,
                                                                                                                                                                                                "time_utc":  1756482123
                                                                                                                                                                                            },
                                                                                                                                                                                 "pubkey_hex":  "deadbeef..."
                                                                                                                                                                             }
                                                                                                                                                               }
                                                                                                                                               }
                                                                                                                              }
                                                                                                     },
                                                                                         "description":  "Latest signed round"
                                                                                     },
                                                                             "404":  {
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/ErrorResponse"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     },
                                                                                         "description":  "No round yet"
                                                                                     }
                                                                         },
                                                           "operationId":  "getLatestSigned"
                                                       }
                                           }
              },
...
    "openapi":  "3.0.3",
    "info":  {
                 "version":  "0.1.0",
                 "title":  "THO Node API"
             }
}